Update fuel station microservice

+++ Fuel-quota-management-system/fuelStation/src/main/java/com/example/fuleStation/repo/FuelStationRepo.java
@@ -4,7 +4,7 @@ import com.example.fuleStation.entity.FuelStation;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface FuelStationRepo extends JpaRepository<FuelStation, Long> {
+public interface FuelStationRepo extends JpaRepository<FuelStation, Integer> {
     @Query(value = "SELECT * FROM fuel_station WHERE id = ?1", nativeQuery = true)
     FuelStation getStationByID(String stationID);
 }
